examples/from_ahmed/Final/requirements.txt
fastapi
uvicorn[standard]
loguru
mysql-connector-python
sqlalchemy
pydantic
databases
